Modularizes example code organization. PR cleanup.
