+++ examples/akka-http-service/pom.xml
@@ -20,7 +20,6 @@
                       http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <properties>
-        <wookiee.version>1.3.17</wookiee.version>
         <scala.version>2.12.10</scala.version>
         <scala.artifact.version>2.12</scala.artifact.version>
         <akka.http.version>10.1.1</akka.http.version>
@@ -43,12 +42,12 @@
     <artifactId>akka-http-service</artifactId>
     <version>1.0-SNAPSHOT</version>
 
-   <!-- <parent>
+    <parent>
         <groupId>com.webtrends</groupId>
         <artifactId>wookiee-akka-http_2.12</artifactId>
         <version>2.0-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
-    </parent>-->
+    </parent>
 
     <repositories>
         <repository>
@@ -59,37 +58,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.webtrends</groupId>
+            <groupId>${parent.groupId}</groupId>
             <artifactId>wookiee-akka-http_2.12</artifactId>
-            <version>2.0-SNAPSHOT</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/../../target/lib/wookiee-akka-http_2.12-2.0-SNAPSHOT.jar</systemPath>
-        </dependency>
-        <dependency>
-            <groupId>com.webtrends</groupId>
-            <artifactId>wookiee-core_${scala.artifact.version}</artifactId>
-            <version>2.0-SNAPSHOT</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/../../lib/wookiee-core_2.12-2.0-SNAPSHOT.jar</systemPath>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-jdk14</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>com.webtrends</groupId>
-            <artifactId>wookiee-test_2.12</artifactId>
-            <version>2.0-SNAPSHOT</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/../../lib/wookiee-test_2.12-2.0-SNAPSHOT.jar</systemPath>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-jdk14</artifactId>
-                </exclusion>
-            </exclusions>
+            <version>${parent.version}</version>
         </dependency>
         <dependency>
             <groupId>com.typesafe.akka</groupId>
