+++ examples/akka-http-service/src/main/scala/com/wookiee/basic/handlers/Responses.scala
@@ -1,0 +1,32 @@
+/*
+ *  Copyright (c) 2020 Oracle and/or its affiliates. All rights reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package com.wookiee.basic.handlers
+
+import akka.http.scaladsl.model.StatusCodes
+import akka.http.scaladsl.server.Directives.complete
+import akka.http.scaladsl.server.Route
+import com.wookiee.basic.handlers.Objects.Message
+
+object Responses {
+  def stringResponse(result: Message): Route = {
+    complete(StatusCodes.OK, result.message)
+  }
+
+  def paramsResponse(result: Seq[String]): Route = {
+    complete(StatusCodes.OK, result.mkString(","))
+  }
+}
