+++ pom.xml
@@ -29,7 +29,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <akka.stream.version>2.5.23</akka.stream.version>
         <akka.http.version>10.1.10</akka.http.version>
-<!--        <wookiee.core.version>1.3.44</wookiee.core.version>-->
+        <wookiee.core.version>2.0.0-RC2</wookiee.core.version>
         <wookiee-metrics.version>2.5.16</wookiee-metrics.version>
         <scala.version>2.12.10</scala.version>
         <scala.artifact.version>2.12</scala.artifact.version>
@@ -66,16 +66,12 @@
         <dependency>
             <groupId>com.webtrends</groupId>
             <artifactId>wookiee-core_${scala.artifact.version}</artifactId>
-            <version>2.0-SNAPSHOT</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/lib/wookiee-core_2.12-2.0-SNAPSHOT.jar</systemPath>
+            <version>${wookiee.core.version}</version>
         </dependency>
         <dependency>
             <groupId>com.webtrends</groupId>
             <artifactId>wookiee-test_2.12</artifactId>
-            <version>2.0-SNAPSHOT</version>
-            <scope>system</scope>
-            <systemPath>${project.basedir}/lib/wookiee-test_2.12-2.0-SNAPSHOT.jar</systemPath>
+            <version>${wookiee.core.version}</version>
         </dependency>
 
         <dependency>
